+++ src/PeakDetect.h
@@ -89,7 +89,7 @@ class PeakDetect
     void extract(const Control& control);
 
     void logPeakStats(
-      const vector<double>& posTrue,
+      const vector<float>& posTrue,
       const string& trainTrue,
       const double speedTrue,
       PeakStats& peakStats);
